refactor: isHeicFormat에서 isHeicFile로 네이밍 변경 #898
