+++ frontend/src/common/utils/heicFile/isHeicFile.ts
